requirements.txt
jsonschema>=4.21.1
PyYAML>=6.0.1
python-dotenv
